Debian: kernel: disable lockdown for secure boot
Debian has this feature for secure boot, while centos stx hasn't:
when secure boot is enabled, whether the sig_enforce is set or not
the unsigned / missigned kernel modules' loading should fail.

To follow debian, we enable lockdown feature for kernel to realize
this function. But now n3000 reset failure is found when lockdown
is enabled. Here are the details: in the process of controller_config
when kernel booted up and the system is being deployed,
puppet calls reset_n3000_fpgas script to run "rsu bmcimg pci_addr"
to reset n3000 fpga. This will fail with secure boot enabled.
The "rsu" tool is from opae-sdk. In the user guide for Intel
FPGA Programmable Acceleration Card N3000:
https://www.intel.com/programmable/technical-pdfs/683040.pdf ,
a note is described:
If your server is set up with secure boot, the OPAE Remote Setup (RSU)
command will not function. This RSU limitation is due to secure boot
having the following limitations:
• Using kexec to start an unsigned kernel image.
• Hibernation and resume from hibernation.
• User-space access to physical memory and I/O ports.
• Module parameters that allow setting memory and I/O port addresses.
• Writing to MSRs through /dev/cpu/*/msr.
• Use of custom ACPI methods and tables.
• ACPI APEI error injection.

Please refer to the lockdown man page:
https://man7.org/linux/man-pages/man7/kernel_lockdown.7.html
We can find that all the limitations mentioned about secure boot in the
n3000 user guide are in the lockdown man page, and they are in fact the
limitations from lockdown.

So we decide to disable lockdown feature in case it causes more gaps
between centos stx and debian stx.

Test Plan:
 - PASS: Do duplex install on wolfpass lab (with N3000 FPGA) with
         secure boot enabled. The reset_n3000_fpgas script runs
         successfully during puppet execution.
 - PASS: The tests are done for both std and rt install.

Closes-Bug: 1998418

Signed-off-by: Li Zhou <li.zhou@windriver.com>
Change-Id: I59cafb3e62f3c6f4c5de5d539cc78d327e5bdcf7
(cherry picked from commit 9292232706187d87e89fd76fb745d58e473853ef)
Jira: CGTS-39668
Signed-off-by: M. Vefa Bicakci <vefa.bicakci@windriver.com>