Intel Quartzville iqvlinux driver v1.2.0.19
The Intel Quartzville (QV) tools is a bundle of network utilities that
support validation, debug, and IEEE Conformance testing of Intel
network adapters; identification and diagnostics of Intel PCI adapters,
collection of properties and assigned resources, for automation; and
an EEPROM programming utility.

The Quartzville tools are proprietary, but the Intel QV Linux (iqvlinux)
driver is licensed as open source (GPL2), and can be freely distributed
as a kernel module.

This commit adds support of Intel Quartzville (QV) tools in StarlingX
kernel, as a facility to users that own a valid Intel QV license, so
they may install QV tools in StarlingX, without changes in the kernel.

The new kernel module is signed, as usual in StarlingX. For this, a
change is necessary in the original source code, to not disable module
signing. This change is done by patch.

Due to iqvlinux 1.2.0.19 is not yet available in sourceforge.net, the
source code tarball shall be hosted in StarlingX CENGN mirror.

Note that the new kernel modules are installed to a directory where
usually only in-tree kernel modules exist. This is done to satisfy the
user-space utilities' assumptions regarding the location of the kernel
module. We have also learned that the user-space utilities' assumption
regarding the kernel module's location may change in the future.

Test Plan:
  - Container image is built successfully on Debian (PASS)
  - iqvlinux driver package is successfully installed (PASS)
  - Run time testing of the QV tools (PASS)

Story: 2010138
Task: 46733

JIRA: CGTS-38777

Signed-off-by: Rogerio Ferraz <RogerioOliveira.Ferraz@windriver.com>
Change-Id: Id12354a68508eaaafcf43db4b62cd4a623638650
(cherry picked from commit 6bc25e523948015fb80eac77bd946a9000955241)